locallibrary/catalog/tests/test_models.py
from django.test import TestCase

# Create your tests here.

from catalog.models import Author

class AuthorModelTest(TestCase):

    @classmethod
    def setUpTestData(cls):
        # Set-up non modified obj used by all test methods
        Author.objects.create(first_name='Big', last_name='Bob')

    def test_first_name_label(self):
        author = Author.objects.get(id=1)
        field_label = author._meta.get_field('first_name').verbose_name
        self.assertEquals(field_label, 'first name')

    def test_date_of_death_label(self):
        author = Author.objects.get(id=1)
        field_label = author._meta.get_field('date_of_death').verbose_name
        self.assertEquals(field_label, 'died')

    def test_first_name_max_length(self):
        author = Author.objects.get(id=1)
        max_length = author._meta.get_field('first_name').max_length
        self.assertEquals(max_length, 100)

    def test_object_name_is_last_name_comma_first_name(self):
        author = Author.objects.get(id=1)
        expected_object_name = '{0} {1}'.format(author.first_name, author.last_name)
        self.assertEquals(expected_object_name, str(author))

    def test_get_absolute_url(self):
        author = Author.objects.get(id=1)
        self.assertEquals(author.get_absolute_url(), '/catalog/author/1')

#====================================================================
# All test methods have same structure
#====================================================================
# 1) author =  Author.objects.get(id=1)
# 1) getting object for testing
#====================================================================
# 2) field_label = author._meta.get_field('first_name').verbose_name
# 2) Getiing meta data field for necessary values
#====================================================================
# 3) Compare values with expected results
# 3) self.asserEquals(field_label, 'first_name')
#====================================================================